refactor: move tool_call_parser resolution and worker_extension_cls into config validators
- Move MODEL_TOOL_CALL_PARSER, resolve_tool_call_parser, WORKER_EXTENSION_CLS to configs/inference.py
- Add auto_resolve_tool_call_parser validator on VLLMConfig
- Set enable_auto_tool_choice in VLLMConfig.to_namespace()
- Add InferenceConfig.to_namespace() that includes worker_extension_cls
- server() now calls config.to_namespace() instead of config.vllm.to_namespace()
